+++ core/src/main/scala/slinky/core/FunctionalComponent.scala
@@ -1,7 +1,7 @@
 package slinky.core
 
 import slinky.readwrite.Reader
-import slinky.core.facade.{React, ReactElement}
+import slinky.core.facade.{React, ReactElement, ReactRef}
 import scala.scalajs.js
 
 import scala.reflect.macros.whitebox
@@ -9,8 +9,7 @@ import scala.language.experimental.macros
 
 import scala.language.implicitConversions
 
-class KeyAddingStage(private[KeyAddingStage] val props: js.Dictionary[js.Any],
-                     private[KeyAddingStage] val constructor: js.Object) {
+final class KeyAddingStage(private val props: js.Dictionary[js.Any], private val constructor: js.Object) {
   def withKey(key: String): ReactElement = {
     props("key") = key
     KeyAddingStage.build(this)
@@ -23,42 +22,88 @@ object KeyAddingStage {
   }
 }
 
-class FunctionalComponent[P] private[core](private[core] val component: js.Object) {
+final class FunctionalComponent[P] private[core](private[core] val component: js.Object) extends AnyVal {
   type Props = P
-
-  def this(fn: P => ReactElement)(implicit fnCompName: FunctionalComponentName) = {
-    this({
-      var ret: js.Function1[js.Object, ReactElement] = null
-      ret = ((obj: js.Object) => {
-        if (obj.hasOwnProperty("__")) {
-          fn(obj.asInstanceOf[js.Dynamic].__.asInstanceOf[P])
-        } else {
-          fn(ret.asInstanceOf[js.Dynamic].__propsReader.asInstanceOf[Reader[P]].read(obj))
-        }
-      })
-
-      if (!scala.scalajs.LinkingInfo.productionMode) {
-        ret.asInstanceOf[js.Dynamic].displayName = fnCompName.name
-      }
-
-      ret
-    }.asInstanceOf[js.Object])
-  }
+  type Result = KeyAddingStage
 
   private[core] def componentWithReader(propsReader: Reader[P]) = {
     component.asInstanceOf[js.Dynamic].__propsReader = propsReader.asInstanceOf[js.Object]
     component
   }
 
-  final def apply(props: P): KeyAddingStage = {
+  @inline final def apply(props: P): KeyAddingStage = {
     new KeyAddingStage(js.Dynamic.literal(
       __ = props.asInstanceOf[js.Any]
     ).asInstanceOf[js.Dictionary[js.Any]], component)
   }
 }
 
+final class FunctionalComponentTakingRef[P, R] private[core](private[core] val component: js.Object) extends AnyVal {
+  type Props = P
+  type Result = KeyAddingStage
+
+  private[core] def componentWithReader(propsReader: Reader[P]) = {
+    component.asInstanceOf[js.Dynamic].__propsReader = propsReader.asInstanceOf[js.Object]
+    component
+  }
+
+  @inline final def apply(props: P): KeyAddingStage = {
+    new KeyAddingStage(js.Dynamic.literal(
+      __ = props.asInstanceOf[js.Any]
+    ).asInstanceOf[js.Dictionary[js.Any]], component)
+  }
+}
+
+final class FunctionalComponentForwardedRef[P, R] private[core](private[core] val component: js.Object) extends AnyVal {
+  type Props = P
+  type Result = KeyAndRefAddingStage[R]
+
+  private[core] def componentWithReader(propsReader: Reader[P]) = {
+    component.asInstanceOf[js.Dynamic].__propsReader = propsReader.asInstanceOf[js.Object]
+    component
+  }
+
+  @inline final def apply(props: P): KeyAndRefAddingStage[R] = {
+    new KeyAndRefAddingStage[R](js.Dynamic.literal(
+      __ = props.asInstanceOf[js.Any]
+    ).asInstanceOf[js.Dictionary[js.Any]], component)
+  }
+}
+
 object FunctionalComponent {
-  def apply[P](fn: P => ReactElement)(implicit name: FunctionalComponentName) = new FunctionalComponent[P](fn)
+  def apply[P](fn: P => ReactElement)(implicit name: FunctionalComponentName) = new FunctionalComponent[P]({
+    var ret: js.Function1[js.Object, ReactElement] = null
+    ret = ((obj: js.Object) => {
+      if (obj.hasOwnProperty("__")) {
+        fn(obj.asInstanceOf[js.Dynamic].__.asInstanceOf[P])
+      } else {
+        fn(ret.asInstanceOf[js.Dynamic].__propsReader.asInstanceOf[Reader[P]].read(obj))
+      }
+    })
+
+    if (!scala.scalajs.LinkingInfo.productionMode) {
+      ret.asInstanceOf[js.Dynamic].displayName = name.name
+    }
+
+    ret
+  }.asInstanceOf[js.Object])
+
+  def apply[P, R](fn: (P, ReactRef[R]) => ReactElement)(implicit name: FunctionalComponentName) = new FunctionalComponentTakingRef[P, R]({
+    var ret: js.Function2[js.Object, ReactRef[R], ReactElement] = null
+    ret = ((obj: js.Object, ref: ReactRef[R]) => {
+      if (obj.hasOwnProperty("__")) {
+        fn(obj.asInstanceOf[js.Dynamic].__.asInstanceOf[P], ref)
+      } else {
+        fn(ret.asInstanceOf[js.Dynamic].__propsReader.asInstanceOf[Reader[P]].read(obj), ref)
+      }
+    })
+
+    if (!scala.scalajs.LinkingInfo.productionMode) {
+      ret.asInstanceOf[js.Dynamic].displayName = name.name
+    }
+
+    ret
+  }.asInstanceOf[js.Object])
 }
 
 final class FunctionalComponentName(val name: String) extends AnyVal
@@ -72,13 +117,15 @@ object FunctionalComponentNameMacros {
 
     // from lihaoyi/sourcecode
     def isSyntheticName(name: String) = {
-      name == "<init>" || (name.startsWith("<local ") && name.endsWith(">"))
+      name == "<init>" || (name.startsWith("<local ") && name.endsWith(">")) || name == "component"
     }
 
     def findNonSyntheticOwner(current: Symbol): Symbol = {
-      if (isSyntheticName(current.name.decodedName.toString)) {
+      if (isSyntheticName(current.name.decodedName.toString.trim)) {
         findNonSyntheticOwner(current.owner)
-      } else current
+      } else {
+        current
+      }
     }
 
     c.Expr(q"new _root_.slinky.core.FunctionalComponentName(${findNonSyntheticOwner(c.internal.enclosingOwner).name.decodedName.toString})")