+++ docs/opt-launcher.js
@@ -2,8 +2,8 @@ var opt = require("./slinky-docs-opt.js");
 
 if (typeof ssr !== 'undefined') {
     module.exports = function render(locals, callback) {
-      callback(null, opt.entrypoint.ssr(locals.path));
+      callback(null, opt.ssr(locals.path));
     };
 } else {
-    opt.entrypoint.hydrate();
+    opt.hydrate();
 }